[tl.py] Display tasks from the last one in
Tasks were added to the list from the beginning of the file onward.
This meant that an old task being worked on recently still showed up
down in the list.

The task list is now ordered from the most recent tasks used which means
that recent tasks will appear higher in the list.

Signed-off-by: Louis Bouchard <louis.bouchard@canonical.com>